Move update/remove polling to DeploymentScanner.
